\name{zigguserDirectC-package}
\alias{zigguserDirectC-package}
\alias{zigguserDirectC}
\docType{package}
\title{\packageTitle{zigguserDirectC}}
\description{\packageDescription{zigguserDirectC}}
\section{Package Content}{\packageIndices{zigguserDirectC}}
\author{\packageAuthor{zigguserDirectC}}
\section{Maintainer}{\packageMaintainer{zigguserDirectC}}
\keyword{package}
